addpath: fix a sleep-deprived bug
Yeah, so last night I added a Show instance to WindowsEnv.Value
(ex-WindowsEnv.VarValue), which would simply call valueString
(ex-varValueString).
Then I replaced every valueString/varValueString with show.
Then I decided that this Show instance was a mistake, and derived an
instance instead.
It obviously messed up all the show VarValue/Value calls.
The lesson to learn is you should remove an instance first, fix all the
call sites, and only then derive it.
The disgusting part is that a few of the last commits are actually
broken, which I hate.